implement Article Entity with manyToOne relation(User)

+++ src/main/java/com/gzella/realworld/business/dto/responses/MultipleArticleResponse.java
@@ -8,6 +8,6 @@ import java.util.List;
 public class MultipleArticleResponse {
 
     private List<ArticleResponse> articles;
-    private int articlesCount;
+    private long articlesCount;
 
 }
